Refactor log messages and type declarations in database engine for improved clarity, consistency, and correctness across journal file and data file operations.

+++ src/database/engine/journalfile.h
@@ -270,7 +270,7 @@ int journalfile_unlink(struct rrdengine_journalfile *journalfile);
  * indicating which on-disk journal files were deleted. This is not a
  * 0 / -errno style status code.
  */
-int journalfile_destroy_unsafe(struct rrdengine_journalfile *journalfile, struct rrdengine_datafile *datafile);
+uint8_t journalfile_destroy_unsafe(struct rrdengine_journalfile *journalfile, struct rrdengine_datafile *datafile);
 int journalfile_create(struct rrdengine_journalfile *journalfile, struct rrdengine_datafile *datafile);
 int journalfile_load(struct rrdengine_instance *ctx, struct rrdengine_journalfile *journalfile,
                      struct rrdengine_datafile *datafile);
